Move external focus point logic to view model
