+++ WordPress/src/main/java/org/wordpress/android/ui/mysite/QuickStartRepository.kt
@@ -4,7 +4,6 @@ import androidx.core.text.HtmlCompat
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.asFlow
-import androidx.lifecycle.distinctUntilChanged
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Job
@@ -20,7 +19,6 @@ import org.wordpress.android.fluxc.generated.SiteActionBuilder
 import org.wordpress.android.fluxc.store.QuickStartStore
 import org.wordpress.android.fluxc.store.QuickStartStore.QuickStartTask
 import org.wordpress.android.fluxc.store.QuickStartStore.QuickStartTask.CREATE_SITE
-import org.wordpress.android.fluxc.store.QuickStartStore.QuickStartTask.FOLLOW_SITE
 import org.wordpress.android.fluxc.store.QuickStartStore.QuickStartTask.PUBLISH_POST
 import org.wordpress.android.fluxc.store.QuickStartStore.QuickStartTask.UPDATE_SITE_TITLE
 import org.wordpress.android.fluxc.store.QuickStartStore.QuickStartTaskType
@@ -39,7 +37,6 @@ import org.wordpress.android.util.EventBusWrapper
 import org.wordpress.android.util.QuickStartUtilsWrapper
 import org.wordpress.android.util.SiteUtils
 import org.wordpress.android.util.analytics.AnalyticsTrackerWrapper
-import org.wordpress.android.util.map
 import org.wordpress.android.viewmodel.Event
 import org.wordpress.android.viewmodel.ResourceProvider
 import javax.inject.Inject
@@ -70,11 +67,6 @@ class QuickStartRepository
     private val _onSnackbar = MutableLiveData<Event<SnackbarMessageHolder>>()
     val onSnackbar = _onSnackbar as LiveData<Event<SnackbarMessageHolder>>
     val activeTask = _activeTask as LiveData<QuickStartTask?>
-
-    val onExternalFocusPointVisibilityChange: LiveData<Event<List<ExternalFocusPointInfo>>> = activeTask
-            .map { getExternalFocusPointInfo(it) }
-            .distinctUntilChanged()
-            .map { Event(it) }
 
     private fun buildQuickStartCategory(siteId: Int, quickStartTaskType: QuickStartTaskType) = QuickStartCategory(
             quickStartTaskType,
@@ -127,12 +119,6 @@ class QuickStartRepository
                     )
                 }
         _onSnackbar.postValue(Event(SnackbarMessageHolder(UiStringText(shortQuickStartMessage))))
-    }
-
-    private fun getExternalFocusPointInfo(task: QuickStartTask?): List<ExternalFocusPointInfo> {
-        // For now, we only do this for the FOLLOW_SITE task.
-        val followSitesTaskFocusPointInfo = ExternalFocusPointInfo(FOLLOW_SITE, task == FOLLOW_SITE)
-        return listOf(followSitesTaskFocusPointInfo)
     }
 
     fun completeTask(task: QuickStartTask) {
